changed unsigned long to size_t for getDimensions
git-svn-id: file:///home/roberto/Git/svn@7428 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 018378834f1cd7fc3a6ff608e229f885037ec83c

+++ trunk/xmipp/applications/programs/convert_image/convert_image_main.cpp
@@ -147,7 +147,7 @@ protected:
             convMode = MD2VOL;
 
             int Xdim, Ydim, Zdim;
-            unsigned long Ndim;
+            size_t Ndim;
             ImgSize(mdIn, Xdim, Ydim, Zdim, Ndim);
             if (Zdim!=1)
                 REPORT_ERROR(ERR_MULTIDIM_DIM,
@@ -159,7 +159,7 @@ protected:
         else if (single_image)
         {
             int Xdim, Ydim, Zdim;
-            unsigned long Ndim;
+            size_t Ndim;
             imTemp.read(fn_in, HEADER);
             imTemp.getDimensions(Xdim,Ydim,Zdim,Ndim);
